media-source/MediaSource-removeSourceBuffer_updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>MediaSource#removeSourceBuffer during the updating attribute is true</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var contents = {'/media/white.webm': 'video/webm; codecs="vorbis,vp8"',
                '/media/white.mp4' : 'video/mp4'};

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
function mediaTest(file, mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            var events = [];
            mediaSource.addEventListener('sourceopen', function(e) {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                sourceBuffer.addEventListener('updatestart', function(e) {
                    events.push('updatestart');
                    t.step(function() {
                        mediaSource.removeSourceBuffer(sourceBuffer);
                        assert_equals(sourceBuffer.updating, false);
                    });
                });
                sourceBuffer.addEventListener('updateend', function(e) {
                    events.push('updateend');
                    t.step(function() {
                        assert_array_equals(events, ['updatestart',
                                                     'abort',
                                                     'updateend']);
                    });
                    t.done();
                });
                sourceBuffer.addEventListener('abort', function(e) {
                    events.push('abort');
                });
                sourceBuffer.appendBuffer(data);
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'MediaSource#removeSourceBuffer (' + mime + ') : ' +
       'If the sourceBuffer.updating attribute equals true, then run the following steps:' +
       '1. Abort the buffer append and stream append loop algorithms if they are running.' +
       '2. Set the sourceBuffer.updating attribute to false.' +
       '3. Queue a task to fire a simple event named abort at sourceBuffer.' +
       '4. Queue a task to fire a simple event named updateend at sourceBuffer.');
}
for(var file in contents) {
    mediaTest(file, contents[file]);
}
</script>
</body>
</html>
